Fix the names of wrongly updated oss package.
