+++ underfs/obs/src/main/resources/META-INF/services/alluxio.underfs.UnderFileSystemFactory
@@ -9,4 +9,4 @@
 # See the NOTICE file distributed with this work for information regarding copyright ownership.
 #
 
-alluxio.underfs.obs.OBSUnderFileSystemFactory
+alluxio.underfs.oss.OBSUnderFileSystemFactory
