+++ underfs/obs/src/test/java/alluxio/underfs/oss/OBSInputStreamTest.java
@@ -9,20 +9,11 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.underfs.obs;
+package alluxio.underfs.oss;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertArrayEquals;
-import static org.junit.Assert.assertEquals;
-import static org.mockito.Matchers.argThat;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.when;
-
-import com.aliyun.oss.OSSClient;
-import com.aliyun.oss.model.GetObjectRequest;
-import com.aliyun.oss.model.OSSObject;
-
+import com.obs.services.ObsClient;
+import com.obs.services.model.GetObjectRequest;
+import com.obs.services.model.S3Object;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -34,18 +25,26 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.Arrays;
 
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
+import static org.mockito.Matchers.argThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.spy;
+import static org.mockito.Mockito.when;
+
 /**
- * Unit tests for the {@link OSSInputStream}.
+ * Unit tests for the {@link OBSInputStream}.
  */
-public class OSSInputStreamTest {
+public class OBSInputStreamTest {
 
   private static final String BUCKET_NAME = "testBucket";
   private static final String OBJECT_KEY = "testObjectKey";
 
-  private OSSInputStream mOssInputStream;
-  private OSSClient mOssClient;
+  private OBSInputStream mOBSInputStream;
+  private ObsClient mObsClient;
   private InputStream[] mInputStreamSpy;
-  private OSSObject[] mOssObject;
+  private S3Object[] mObjects;
 
   /**
    * The exception expected to be thrown.
@@ -55,59 +54,59 @@ public class OSSInputStreamTest {
 
   @Before
   public void setUp() throws IOException {
-    mOssClient = mock(OSSClient.class);
+    mObsClient = mock(ObsClient.class);
 
     byte[] input = new byte[] {1, 2, 3};
-    mOssObject = new OSSObject[input.length];
+    mObjects = new S3Object[input.length];
     mInputStreamSpy = new InputStream[input.length];
     for (int i = 0; i < input.length; ++i) {
       final long pos = (long) i;
-      mOssObject[i] = mock(OSSObject.class);
-      when(mOssClient.getObject(argThat(new ArgumentMatcher<GetObjectRequest>() {
+      mObjects[i] = mock(S3Object.class);
+      when(mObsClient.getObject(argThat(new ArgumentMatcher<GetObjectRequest>() {
         @Override
         public boolean matches(Object argument) {
-          if (argument instanceof  GetObjectRequest) {
-            return ((GetObjectRequest) argument).getRange()[0] == pos;
+          if (argument instanceof GetObjectRequest) {
+            return ((GetObjectRequest) argument).getRangeStart() == pos;
           }
           return false;
         }
-      }))).thenReturn(mOssObject[i]);
+      }))).thenReturn(mObjects[i]);
       byte[] mockInput = Arrays.copyOfRange(input, i, input.length);
       mInputStreamSpy[i] = spy(new ByteArrayInputStream(mockInput));
-      when(mOssObject[i].getObjectContent()).thenReturn(mInputStreamSpy[i]);
+      when(mObjects[i].getObjectContent()).thenReturn(mInputStreamSpy[i]);
     }
-    mOssInputStream = new OSSInputStream(BUCKET_NAME, OBJECT_KEY, mOssClient);
+    mOBSInputStream = new OBSInputStream(BUCKET_NAME, OBJECT_KEY, mObsClient);
   }
 
   @Test
   public void close() throws IOException {
-    mOssInputStream.close();
+    mOBSInputStream.close();
 
     mExceptionRule.expect(IOException.class);
     mExceptionRule.expectMessage(is("Stream closed"));
-    mOssInputStream.read();
+    mOBSInputStream.read();
   }
 
   @Test
   public void readInt() throws IOException {
-    assertEquals(1, mOssInputStream.read());
-    assertEquals(2, mOssInputStream.read());
-    assertEquals(3, mOssInputStream.read());
+    assertEquals(1, mOBSInputStream.read());
+    assertEquals(2, mOBSInputStream.read());
+    assertEquals(3, mOBSInputStream.read());
   }
 
   @Test
   public void readByteArray() throws IOException {
     byte[] bytes = new byte[3];
-    int readCount = mOssInputStream.read(bytes, 0, 3);
+    int readCount = mOBSInputStream.read(bytes, 0, 3);
     assertEquals(3, readCount);
     assertArrayEquals(new byte[] {1, 2, 3}, bytes);
   }
 
   @Test
   public void skip() throws IOException {
-    assertEquals(1, mOssInputStream.read());
-    mOssInputStream.skip(1);
-    assertEquals(3, mOssInputStream.read());
+    assertEquals(1, mOBSInputStream.read());
+    mOBSInputStream.skip(1);
+    assertEquals(3, mOBSInputStream.read());
   }
 
 }
