+++ underfs/obs/src/test/java/alluxio/underfs/oss/OBSOutputStreamTest.java
@@ -9,11 +9,11 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.underfs.obs;
+package alluxio.underfs.oss;
 
-import com.aliyun.oss.OSSClient;
-import com.aliyun.oss.OSSException;
-import com.aliyun.oss.model.ObjectMetadata;
+import com.obs.services.ObsClient;
+import com.obs.services.exception.ObsException;
+import com.obs.services.model.ObjectMetadata;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -35,11 +35,11 @@ import java.io.OutputStream;
 import java.security.DigestOutputStream;
 
 /**
- * Unit tests for the {@link OSSOutputStream}.
+ * Unit tests for the {@link OBSOutputStream}.
  */
 @RunWith(PowerMockRunner.class)
-public class OSSOutputStreamTest {
-  private OSSClient mOssClient;
+public class OBSOutputStreamTest {
+  private ObsClient mObsClient;
   private File mFile;
   private BufferedOutputStream mLocalOutputStream;
 
@@ -54,7 +54,7 @@ public class OSSOutputStreamTest {
    */
   @Before
   public void before() throws Exception {
-    mOssClient = Mockito.mock(OSSClient.class);
+    mObsClient = Mockito.mock(ObsClient.class);
     mFile = Mockito.mock(File.class);
     mLocalOutputStream = Mockito.mock(BufferedOutputStream.class);
   }
@@ -63,7 +63,7 @@ public class OSSOutputStreamTest {
    * Tests to ensure IOException is thrown if {@link FileOutputStream()} throws an IOException.
    */
   @Test
-  @PrepareForTest(OSSOutputStream.class)
+  @PrepareForTest(OBSOutputStream.class)
   public void testConstructor() throws Exception {
     PowerMockito.whenNew(File.class).withArguments(Mockito.anyString()).thenReturn(mFile);
     String errorMessage = "protocol doesn't support output";
@@ -71,37 +71,37 @@ public class OSSOutputStreamTest {
             .thenThrow(new IOException(errorMessage));
     mThrown.expect(IOException.class);
     mThrown.expectMessage(errorMessage);
-    new OSSOutputStream("testBucketName", "testKey", mOssClient).close();
+    new OBSOutputStream("testBucketName", "testKey", mObsClient).close();
   }
 
   /**
-   * Tests to ensure {@link OSSOutputStream#write(int)} calls {@link OutputStream#write(int)}.
+   * Tests to ensure {@link OBSOutputStream#write(int)} calls {@link OutputStream#write(int)}.
    */
   @Test
-  @PrepareForTest(OSSOutputStream.class)
+  @PrepareForTest(OBSOutputStream.class)
   public void testWrite1() throws Exception {
     PowerMockito.whenNew(BufferedOutputStream.class)
             .withArguments(Mockito.any(DigestOutputStream.class)).thenReturn(mLocalOutputStream);
     PowerMockito.whenNew(BufferedOutputStream.class)
             .withArguments(Mockito.any(FileOutputStream.class)).thenReturn(mLocalOutputStream);
-    OSSOutputStream stream = new OSSOutputStream("testBucketName", "testKey", mOssClient);
+    OBSOutputStream stream = new OBSOutputStream("testBucketName", "testKey", mObsClient);
     stream.write(1);
     stream.close();
     Mockito.verify(mLocalOutputStream).write(1);
   }
 
   /**
-   * Tests to ensure {@link OSSOutputStream#write(byte[], int, int)} calls
+   * Tests to ensure {@link OBSOutputStream#write(byte[], int, int)} calls
    * {@link OutputStream#write(byte[], int, int)} .
    */
   @Test
-  @PrepareForTest(OSSOutputStream.class)
+  @PrepareForTest(OBSOutputStream.class)
   public void testWrite2() throws Exception {
     PowerMockito.whenNew(BufferedOutputStream.class)
             .withArguments(Mockito.any(DigestOutputStream.class)).thenReturn(mLocalOutputStream);
     PowerMockito.whenNew(BufferedOutputStream.class)
             .withArguments(Mockito.any(FileOutputStream.class)).thenReturn(mLocalOutputStream);
-    OSSOutputStream stream = new OSSOutputStream("testBucketName", "testKey", mOssClient);
+    OBSOutputStream stream = new OBSOutputStream("testBucketName", "testKey", mObsClient);
     byte[] b = new byte[1];
     stream.write(b, 0, 1);
     stream.close();
@@ -109,16 +109,16 @@ public class OSSOutputStreamTest {
   }
 
   /**
-   * Tests to ensure {@link OSSOutputStream#write(byte[])} calls {@link OutputStream#write(byte[])}.
+   * Tests to ensure {@link OBSOutputStream#write(byte[])} calls {@link OutputStream#write(byte[])}.
    */
   @Test
-  @PrepareForTest(OSSOutputStream.class)
+  @PrepareForTest(OBSOutputStream.class)
   public void testWrite3() throws Exception {
     PowerMockito.whenNew(BufferedOutputStream.class)
             .withArguments(Mockito.any(DigestOutputStream.class)).thenReturn(mLocalOutputStream);
     PowerMockito.whenNew(BufferedOutputStream.class)
             .withArguments(Mockito.any(FileOutputStream.class)).thenReturn(mLocalOutputStream);
-    OSSOutputStream stream = new OSSOutputStream("testBucketName", "testKey", mOssClient);
+    OBSOutputStream stream = new OBSOutputStream("testBucketName", "testKey", mObsClient);
     byte[] b = new byte[1];
     stream.write(b);
     stream.close();
@@ -127,21 +127,21 @@ public class OSSOutputStreamTest {
 
   /**
    * Tests to ensure IOException is thrown if
-   * {@link OSSClient#putObject(String, String, InputStream, ObjectMetadata)} throws an
-   * OSSException.
+   * {@link ObsClient#putObject(String, String, InputStream, ObjectMetadata)} throws an
+   * {@link com.obs.services.exception.ObsException}.
    */
   @Test
-  @PrepareForTest(OSSOutputStream.class)
+  @PrepareForTest(OBSOutputStream.class)
   public void testCloseError() throws Exception {
     String errorMessage = "Invoke the createEmptyObject method error.";
     BufferedInputStream inputStream = PowerMockito.mock(BufferedInputStream.class);
     PowerMockito.whenNew(BufferedInputStream.class)
             .withArguments(Mockito.any(FileInputStream.class)).thenReturn(inputStream);
     PowerMockito
-            .when(mOssClient.putObject(Mockito.anyString(), Mockito.anyString(),
+            .when(mObsClient.putObject(Mockito.anyString(), Mockito.anyString(),
                     Mockito.any(InputStream.class), Mockito.any(ObjectMetadata.class)))
-            .thenThrow(new OSSException(errorMessage));
-    OSSOutputStream stream = new OSSOutputStream("testBucketName", "testKey", mOssClient);
+            .thenThrow(new ObsException(errorMessage));
+    OBSOutputStream stream = new OBSOutputStream("testBucketName", "testKey", mObsClient);
     mThrown.expect(IOException.class);
     mThrown.expectMessage(errorMessage);
     stream.close();
@@ -151,7 +151,7 @@ public class OSSOutputStreamTest {
    * Tests to ensure {@link File#delete()} is called when close the stream.
    */
   @Test
-  @PrepareForTest(OSSOutputStream.class)
+  @PrepareForTest(OBSOutputStream.class)
   public void testCloseSuccess() throws Exception {
     PowerMockito.whenNew(File.class).withArguments(Mockito.anyString()).thenReturn(mFile);
     FileOutputStream outputStream = PowerMockito.mock(FileOutputStream.class);
@@ -159,20 +159,20 @@ public class OSSOutputStreamTest {
     FileInputStream inputStream = PowerMockito.mock(FileInputStream.class);
     PowerMockito.whenNew(FileInputStream.class).withArguments(mFile).thenReturn(inputStream);
 
-    OSSOutputStream stream = new OSSOutputStream("testBucketName", "testKey", mOssClient);
+    OBSOutputStream stream = new OBSOutputStream("testBucketName", "testKey", mObsClient);
     stream.close();
     Mockito.verify(mFile).delete();
   }
 
   /**
-   * Tests to ensure {@link OSSOutputStream#flush()} calls {@link OutputStream#flush()}.
+   * Tests to ensure {@link OBSOutputStream#flush()} calls {@link OutputStream#flush()}.
    */
   @Test
-  @PrepareForTest(OSSOutputStream.class)
+  @PrepareForTest(OBSOutputStream.class)
   public void testFlush() throws Exception {
     PowerMockito.whenNew(BufferedOutputStream.class)
             .withArguments(Mockito.any(DigestOutputStream.class)).thenReturn(mLocalOutputStream);
-    OSSOutputStream stream = new OSSOutputStream("testBucketName", "testKey", mOssClient);
+    OBSOutputStream stream = new OBSOutputStream("testBucketName", "testKey", mObsClient);
     stream.flush();
     stream.close();
     Mockito.verify(mLocalOutputStream).flush();
