+++ underfs/oss/src/main/java/alluxio/underfs/oss/OSSInputStream.java
@@ -9,7 +9,7 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.underfs.obs;
+package alluxio.underfs.oss;
 
 import alluxio.underfs.MultiRangeObjectInputStream;
