+++ underfs/oss/src/test/java/alluxio/underfs/oss/OSSInputStreamTest.java
@@ -9,11 +9,11 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.underfs.obs;
+package alluxio.underfs.oss;
 
-import com.obs.services.ObsClient;
-import com.obs.services.model.GetObjectRequest;
-import com.obs.services.model.S3Object;
+import com.aliyun.oss.OSSClient;
+import com.aliyun.oss.model.GetObjectRequest;
+import com.aliyun.oss.model.OSSObject;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -34,17 +34,17 @@ import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.when;
 
 /**
- * Unit tests for the {@link OBSInputStream}.
+ * Unit tests for the {@link OSSInputStream}.
  */
-public class OBSInputStreamTest {
+public class OSSInputStreamTest {
 
   private static final String BUCKET_NAME = "testBucket";
   private static final String OBJECT_KEY = "testObjectKey";
 
-  private OBSInputStream mOBSInputStream;
-  private ObsClient mObsClient;
+  private OSSInputStream mOssInputStream;
+  private OSSClient mOssClient;
   private InputStream[] mInputStreamSpy;
-  private S3Object[] mObjects;
+  private OSSObject[] mOssObject;
 
   /**
    * The exception expected to be thrown.
@@ -54,59 +54,59 @@ public class OBSInputStreamTest {
 
   @Before
   public void setUp() throws IOException {
-    mObsClient = mock(ObsClient.class);
+    mOssClient = mock(OSSClient.class);
 
     byte[] input = new byte[] {1, 2, 3};
-    mObjects = new S3Object[input.length];
+    mOssObject = new OSSObject[input.length];
     mInputStreamSpy = new InputStream[input.length];
     for (int i = 0; i < input.length; ++i) {
       final long pos = (long) i;
-      mObjects[i] = mock(S3Object.class);
-      when(mObsClient.getObject(argThat(new ArgumentMatcher<GetObjectRequest>() {
+      mOssObject[i] = mock(OSSObject.class);
+      when(mOssClient.getObject(argThat(new ArgumentMatcher<GetObjectRequest>() {
         @Override
         public boolean matches(Object argument) {
-          if (argument instanceof GetObjectRequest) {
-            return ((GetObjectRequest) argument).getRangeStart() == pos;
+          if (argument instanceof  GetObjectRequest) {
+            return ((GetObjectRequest) argument).getRange()[0] == pos;
           }
           return false;
         }
-      }))).thenReturn(mObjects[i]);
+      }))).thenReturn(mOssObject[i]);
       byte[] mockInput = Arrays.copyOfRange(input, i, input.length);
       mInputStreamSpy[i] = spy(new ByteArrayInputStream(mockInput));
-      when(mObjects[i].getObjectContent()).thenReturn(mInputStreamSpy[i]);
+      when(mOssObject[i].getObjectContent()).thenReturn(mInputStreamSpy[i]);
     }
-    mOBSInputStream = new OBSInputStream(BUCKET_NAME, OBJECT_KEY, mObsClient);
+    mOssInputStream = new OSSInputStream(BUCKET_NAME, OBJECT_KEY, mOssClient);
   }
 
   @Test
   public void close() throws IOException {
-    mOBSInputStream.close();
+    mOssInputStream.close();
 
     mExceptionRule.expect(IOException.class);
     mExceptionRule.expectMessage(is("Stream closed"));
-    mOBSInputStream.read();
+    mOssInputStream.read();
   }
 
   @Test
   public void readInt() throws IOException {
-    assertEquals(1, mOBSInputStream.read());
-    assertEquals(2, mOBSInputStream.read());
-    assertEquals(3, mOBSInputStream.read());
+    assertEquals(1, mOssInputStream.read());
+    assertEquals(2, mOssInputStream.read());
+    assertEquals(3, mOssInputStream.read());
   }
 
   @Test
   public void readByteArray() throws IOException {
     byte[] bytes = new byte[3];
-    int readCount = mOBSInputStream.read(bytes, 0, 3);
+    int readCount = mOssInputStream.read(bytes, 0, 3);
     assertEquals(3, readCount);
     assertArrayEquals(new byte[] {1, 2, 3}, bytes);
   }
 
   @Test
   public void skip() throws IOException {
-    assertEquals(1, mOBSInputStream.read());
-    mOBSInputStream.skip(1);
-    assertEquals(3, mOBSInputStream.read());
+    assertEquals(1, mOssInputStream.read());
+    mOssInputStream.skip(1);
+    assertEquals(3, mOssInputStream.read());
   }
 
 }
